Finalizando

+++ src/AgGraph.h
@@ -2,6 +2,6 @@
 #define AGGRAPH_H_
 
 int evolve(Gene *queenAg, int generationBestIndex);
-int evolveGraphAg(Gene *queenAg);
+long long int evolveGraphAg(Gene *queenAg);
 
 #endif
